akdirabiliterre/requirement.txt
Babel==2.5.3
Django==1.11
django-countries==5.3
django-leaflet==0.24.0
django-phonenumber-field==2.0.0
django-summernote==0.8.8.6
django-taggit==0.22.2
phonenumberslite==8.9.6
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
